autogen.sh: easy/automated regenaration of build stuff (for packagers)
The currently used script regenerate_all.sh is far from being optimal,
and should _only_ be used by maintainer, *if configure.ac needs to be
updated.

OTOH, packagers sometimes need to regenerate the whole autotools stuff,
and this should work fully automatically. (in general, if regeneration
on the build environment leads to any problem, this should be considered
a bug, which should be fixed cleanly instead of relying on developers
to provide pre-generated autotools-files).

This script also allows specifying the actual autotools commands via
environment, in case of packaging system needs to point to specific
versions. (eg. pretty usual on Gentoo)